rules/community/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
aiosqlite==0.20.0
pefile==2024.8.26
anthropic==0.39.0
markdown==3.7
oletools==0.60.2
pyelftools==0.31
yara-python==4.5.1
ppdeep==20200505
py-tlsh==4.7.2
